Handle empty events array
It's common to pre-emptively call `off` with an event and handler prior
to any handlers being registered. This is a safeguard to ensure that the
same handler isn't attached multiple times. In this scenario, the events
array isn't checked for existence prior to the accessor call (for the
given selector). Thus a TypeError is raised.

Closes #31